Add a script to setup and configure
We use `rcm` to setup the dotfiles in place. Ubuntu 19.04 and up has it
in the repositories already. We always need to ensure we update on
setup, as it's unpredictable whether or not this has been recently run.
We also want to fail on setup error, because trying to continue ends up
with a weird state.