Add overrides for GitHub Codespaces
Codespaces uses a significantly different `dotfiles` path, because it's
located in `/workspaces/.codespaces/.persistedshare/dotfiles`. We need
to override the assumed directory for `rcm`, and we need to delete the
existing `zsh` configuration because otherwise they clash. We could
force to overwrite, but this is more deliberate.

To check that we're running in a Codespace, we check for the presence of
`$CODESPACES`, which is unlikely to be set elsewhere.

https://docs.github.com/en/codespaces/troubleshooting/troubleshooting-personalization-for-codespaces